Estrutura inicial p/  nova página de estatisticas sobre vendas

+++ react/src/pages/produtos/cadastro/cadastro.tsx
@@ -42,7 +42,7 @@ export default function CadastroProduto() {
           label="Preço"
           placeholder="12.99"
           onChange={(e) => {
-            setPreco(parseInt(e.target.value));
+            setPreco(parseFloat(e.target.value));
           }}
         />
         <TextField
